Extract recipe benchmark artifact helpers
